alsa_pcm: extract ALSA mmap FFI + AlsaMmapCtx into shared module
Same code, new home. lib.rs no longer carries the snd_pcm_* extern
block, AlsaHandle, MmapAudioFormat, or AlsaMmapCtx in line; they now
live in alsa_pcm.rs so V2 native_transport's upcoming ALSA output
backend can pull them in without circular deps.
